Automate compilation
Given we're developing on macOS we need need to
cross-compile the program for the Linux kernel.
Instead of playing aroung with out shell's
environment variables, we'd rather pack it all
up within a single Makefile. That way everything
is cleaner and we don't have to `unset` `GOOS`
every time we just want to locally test the code
with `go test`...